added chat service and front-end

+++ src/main/java/websockets/chat/ChatWebSocket.java
@@ -3,10 +3,7 @@ package websockets.chat;
 import dao.UserDAO;
 import models.User;
 import org.eclipse.jetty.websocket.api.Session;
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketClose;
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketConnect;
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketMessage;
-import org.eclipse.jetty.websocket.api.annotations.WebSocket;
+import org.eclipse.jetty.websocket.api.annotations.*;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -23,16 +20,20 @@ public class ChatWebSocket {
     private UserDAO userDAO;
     private User user;
     private boolean adminMode;
+    private boolean loginError = false;
     private ArrayList<Pattern> patterns = new ArrayList<>();
-    private final Pattern banCommandPattern = Pattern.compile("^\\/auth ([\\S]+).*");
 
     String login;
 
     public ChatWebSocket(ChatService chatService, UserDAO userDAO, String login) {
+        if (login == null) {
+            loginError = true;
+            return;
+        }
         this.chatService = chatService;
         this.userDAO = userDAO;
-        user = this.userDAO.findByLogin(login);
-        this.login=login;
+        this.user = this.userDAO.findByLogin(login);
+        this.login = login;
         adminMode = user.getLogin().equals("root");
 
         patterns.add(Pattern.compile("^/ban ([\\S]+).*"));
@@ -42,42 +43,46 @@ public class ChatWebSocket {
 
     @OnWebSocketConnect
     public void onOpen(Session session) {
+        if (loginError) {
+            session.close();
+            return;
+        }
         chatService.add(this);
         this.session = session;
         chatService.history.forEach(this::sendString);
-        chatService.sendMessage(currentTime() + user.getLogin() + " joined!",false);
+        chatService.sendMessage(currentTime() + user.getLogin() + " joined!", false);
         chatService.updateOnline();
     }
 
     @OnWebSocketMessage
     public void onMessage(String data) {
         int patternIndex = patternValidate(data);
-        if(adminMode && patternIndex>=0){
+        if (adminMode && patternIndex >= 0) {
             Matcher matcher = patterns.get(patternIndex).matcher(data);
-            matcher.find();
-            switch (patternIndex){
+            switch (patternIndex) {
                 case 0:
-                    userDAO.banControlByLogin(matcher.group(1),true);
-                    chatService.sendMessage(currentTime() + "User "+matcher.group(1)+" was banned!",false);
+                    userDAO.banControlByLogin(matcher.group(1), true);
+                    chatService.sendMessage(currentTime() + "User " + matcher.group(1) + " was banned!", false);
                     break;
                 case 1:
-                    userDAO.banControlByLogin(matcher.group(1),false);
-                    chatService.sendMessage(currentTime() + "User "+matcher.group(1)+" was unbanned!",false);
+                    userDAO.banControlByLogin(matcher.group(1), false);
+                    chatService.sendMessage(currentTime() + "User " + matcher.group(1) + " was unbanned!", false);
                     break;
                 case 2:
-                    chatService.sendMessage("/clear",true);
+                    chatService.sendMessage("/clear", true);
                     chatService.history.clear();
                     break;
             }
-        } else{
-            chatService.sendMessage(currentTime() + data,false);
+        } else {
+            chatService.sendMessage(currentTime() + user.getLogin() + ": " + data, false);
         }
     }
 
     @OnWebSocketClose
     public void onClose(int statusCode, String reason) {
+        if (loginError) return;
         chatService.remove(this);
-        chatService.sendMessage(currentTime() + user.getLogin() + " left!",false);
+        chatService.sendMessage(currentTime() + user.getLogin() + " left!", false);
         chatService.updateOnline();
     }
 
@@ -94,9 +99,9 @@ public class ChatWebSocket {
         return time + "\t\t";
     }
 
-    private int patternValidate(String data){
-        for(int i=0;i<patterns.size();i++){
-            if(patterns.get(i).matcher(data).matches()){
+    private int patternValidate(String data) {
+        for (int i = 0; i < patterns.size(); i++) {
+            if (patterns.get(i).matcher(data).matches()) {
                 return i;
             }
         }
